Move VRAD music into postcompiler package
